Created consolidated compendium
As of v1.2.2 of Fight Club and v1.1.1. of Game Master, separate files
for items, spells, and bestiary are no longer necessary.  I simply
consolidated the existing files into a single file for the core books.
This simplifies the importation process down to one step. Single files
could also be created for the items, spells, and creatures associated
with each adventure.